sutt

+++ code/modules/events/moderate/herb_wither.dm
@@ -15,6 +15,16 @@
 
 /datum/round_event/herb_wither/start()
 	. = ..()
+	if(!length(GLOB.herb_locations))
+		return
 	for(var/i = 1 to rand(1, 5))
+		if(!length(GLOB.herb_locations))
+			break
 		var/obj/structure/flora/roguegrass/herb/herb = pick(GLOB.herb_locations)
-		qdel(herb)
+		if(!QDELETED(herb))
+			qdel(herb)
+
+/datum/round_event_control/herb_wither/canSpawnEvent(players_amt, gamemode, fake_check)
+	if(!length(GLOB.herb_locations))
+		return FALSE
+	return ..()
